Minor modifications in kernel implementation

+++ src/kernel.h
@@ -6,7 +6,7 @@
 class Kernel {
     private:
         /** @brief The values of the Kernel. **/
-        std::vector<std::vector<float>> kernelData;
+        std::vector<std::vector<double>> kernelData;
 
     public:
         /** 
@@ -15,7 +15,7 @@ class Kernel {
          * @warning The input matrix must not be empty, have even dimensions, or
          *          different row-dimensions. Otherwise, std::invalid_argument is thrown.
          */
-        Kernel(const std::vector<std::vector<float>>& kernelData);
+        Kernel(const std::vector<std::vector<double>>& kernelData);
 
         /** @brief Copy-construct the Kernel. **/
         Kernel(const Kernel& kernel);
@@ -25,10 +25,10 @@ class Kernel {
          *
          * @returns The factor used for the normalization.
          */
-        float normalize();
+        double normalize();
 
         /** @brief Returns the specified row of the Kernel. **/
-        std::vector<float> operator[](const size_t index) const;
+        std::vector<double> operator[](const size_t index) const;
         
         /** @brief Returns the height of the Kernel. **/
         size_t getHeight() const;
@@ -52,9 +52,9 @@ class GaussianKernel : public Kernel {
          * @warning width and height must be odd and the variance positive. 
          *          std::invalid_argument is thrown otherwise.
          */
-        std::vector<std::vector<float>> getGaussianData(const size_t width, 
-                                                        const size_t height, 
-                                                        const float variance) const;
+        std::vector<std::vector<double>> getGaussianData(const size_t width, 
+                                                         const size_t height, 
+                                                         const double variance) const;
 
     public:
         /**
@@ -63,5 +63,5 @@ class GaussianKernel : public Kernel {
          *
          * Calls getGaussianData(). All requirements apply.
          */
-        GaussianKernel(const size_t width, const size_t height, const float variance);
+        GaussianKernel(const size_t width, const size_t height, const double variance);
 };
